requirements.txt

streamlit
torch
torchvision
Pillow
numpy
qrcode
